Create a template to fill the credit author statement

Create a template to fill the credit author statement

create_template(authors, roles = NULL, file = NULL)

Arguments

authors

A character vector with all the authors to be included in the statement

roles

A character vector with the roles to be included in the statement. If NULL, it uses all the roles defined in the CRediT author statement

file

If a path is provided, the template is saved as a csv for excel

Value

A dataframe with a row for each authors and a column for each role, filled with zeros.

Details

The dataframe can be exported to a csv to be edited manually.

Examples

create_template(authors = c("Josep Maria", "Jane Doe"))
#>       Authors Conceptualization Methodology Software Validation Formal analysis
#> 1 Josep Maria                 0           0        0          0               0
#> 2    Jane Doe                 0           0        0          0               0
#>   Investigation Resources Data curation Writing - Original Draft
#> 1             0         0             0                        0
#> 2             0         0             0                        0
#>   Writing - Review & Editing Visualization Supervision Project administration
#> 1                          0             0           0                      0
#> 2                          0             0           0                      0
#>   Funding acquisition
#> 1                   0
#> 2                   0